major bug fixes + debugging
the bot was placing pieces faster than it captures the grid and queue, so it was playing on a wrong grid.

Added more todo and implemented an idea of simulating all the pieces at once instead of recursively (this method has its issues)

one of the major bugs was not updating the curPiece (redeclared it in main), added more debugging junk, will remove soon.